Initial state is read from an XML file instead of hard coded. Bugfixes.
